+++ includes/menus/class.primary-menu-walker.php
@@ -12,7 +12,7 @@ class Ecologie_Primary_Nav_Menu_Walker extends Walker_Nav_Menu {
 	 * @access public
 	 *
 	 * @param string $output The markup to be rendered.
-	 * @param integer $depth Depth within menu.
+	 * @param int $depth Depth within menu.
 	 * @param array $args Further arguments.
 	 */
 	public function start_lvl( &$output, $depth = 0, $args = array() ) {
@@ -27,9 +27,9 @@ class Ecologie_Primary_Nav_Menu_Walker extends Walker_Nav_Menu {
 	 *
 	 * @param string $output The markup to be rendered.
 	 * @param object $item Menu item.
-	 * @param integer $depth Depth within menu.
+	 * @param int $depth Depth within menu.
 	 * @param array $args Further arguments.
-	 * @param integer $id Menu item ID.
+	 * @param int $id Menu item ID.
 	 */
 	public function start_el( &$output, $item, $depth = 0, $args = array(), $id = 0 ) {
 		$indent = ( $depth ) ? str_repeat( "\t", $depth ) : '';