+++ changelog.md
@@ -4,7 +4,7 @@
 * Contact Widget: Calling codes now visible in dropdown options
 * Navigation Menus: No bullets/separators if no menu is yet set for location
 * Site title and description in header and footer are now limited to the first 40 and 120 characters respectively
-* Code cleanup
+* Code cleanup and documentation changes
 
 # 0.9 (27-12-2017)