+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Contact Shortcode: Option to enable/disable captcha
 * Comment navigation arrows
+* Default buttons gone greenish
 * Mailchimp Subscribe Widget: Fixed regression with interaction
 * Contact Widget: Country codes now fetched through restcountries.eu
 * Contact Widget: Calling codes now visible in dropdown options