+++ changelog.md
@@ -1,7 +1,7 @@
 # 0.9.1 ([Release date])
 
 * Back to top button
-* Contact Shortcode: Invisible Recaptcha replacing arithmetic captcha
+* Contact Shortcode: Invisible reCAPTCHA replacing arithmetic captcha
 * Contact Shortcode: Redirect URL field in Customiser
 * Contact Shortcode: Upon authentication with Google, the link to authenticate is changed to text reading "Authenticated!"
 * Contact Shortcode: Option to enable/disable captcha