+++ changelog.md
@@ -21,6 +21,7 @@
 * Navigation Menus: No bullets/separators if no menu is yet set for location
 * Site title and description in header and footer are now limited to the first 40 and 120 characters respectively
 * Shortened text shown in header for development mode to just " | Dev mode"
+* Footer: Top part is hidden in case no widget area contains widgets
 * Code cleanup and documentation changes
 
 # 0.9 (27-12-2017)